+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.0.5 (2021-06-16)
+
+### Features
+
+- ASYNCIO support play
+- Record default 5 seconds and more selected.
+
+### Issues
+
+- Record async not working
+
+
 # 0.0.2 (2021-06-15)
 
 ### Features